requirements.txt
streamlit==1.44.1
firebase-admin==6.7.0
requests==2.32.3
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.2.5
pandas==2.2.3
nltk==3.9.1
pdfplumber==0.11.5
python-docx==1.1.2
